Fix validation function to use actual data
The validation functions were only referring to the form elements and
then fell back on a hard-coded value. This doesn't always work, for
example when a change to the flow in a project is reverted and the whole
UI reloads.

As the documentation[^1] recommends this change now uses `this` to read
the stored value in case the form value isn't available.

[^1]: https://nodered.org/docs/creating-nodes/properties#property-validation